standingIndex: Optional[float] -> Optional[NonNegativeInt]
Nothing ever reads the magnitude -- both solver and evaluator only compare
it to establish an order -- and no value in the corpus was ever fractional.
cTORS's own TrainGoal already types it `const int` internally and silently
truncates the wire double at construction; this makes the schema honest
about what was already true downstream. Also drops two now-stale TODOs
superseded by the "Standing order" decision in unified-schema-design.md
(standingIndex's presence on IncomingTrain was already resolved: yes, both).

Regenerated schema/schema_scenario.json and schema_combined.json via
scripts/export_schema.py; only the standingIndex fragments changed.

+++ src/scenario.py
@@ -205,7 +205,7 @@ class ScenarioGenerator:
         self.scenario_train_unit_types.append(train_unit_type)
 
     @staticmethod
-    def create_train(side_track_part: int, track_part: int, time: int, id: str, members: List[TrainUnit], can_depart_from_any_track: bool = True, standing_index: Optional[float] = None, minimum_duration: str = "60")->Train:
+    def create_train(side_track_part: int, track_part: int, time: int, id: str, members: List[TrainUnit], can_depart_from_any_track: bool = True, standing_index: Optional[int] = None, minimum_duration: str = "60")->Train:
         """_summary_
         Method used to create train objects that are added either as an in- or an out-going train.
 
@@ -218,7 +218,7 @@ class ScenarioGenerator:
             id (str): unique identifier of the Train
             members (List[TrainUnit]): train units in the train
             can_depart_from_any_track (bool): For outstanding trains: set to true to allow departures from any track, instead of just the parking_track_part (TORS required, not used)
-            standing_index (Optional[float]): if train is in- or outstanding and there are multiple trains on one track, use this to determine the index of the train on the track, with lower indices at the A-side of the track. Leave unset if there is no preference (or only one train stands on the track).
+            standing_index (Optional[int]): if train is in- or outstanding and there are multiple trains on one track, use this to determine the index of the train on the track, with lower indices at the A-side of the track. Leave unset if there is no preference (or only one train stands on the track).
             minimum_duration (str): minimum duration on the track part where the train arrives/departs
 
         Returns:
